+++ inc/checkout-coupon-codes.php
@@ -25,24 +25,29 @@ class FluidCheckout_CouponCodes extends FluidCheckout {
 		// Body Class
 		add_filter( 'body_class', array( $this, 'add_body_class' ) );
 
-		// Checkout Coupon Notice
-		remove_action( 'woocommerce_before_checkout_form', 'woocommerce_checkout_coupon_form', 10 );
+		// Enqueue
+		add_action( 'wp_enqueue_scripts', array( $this, 'register_assets' ), 5 );
+		add_action( 'wp_enqueue_scripts', array( $this, 'enqueue_assets' ), 10 );
 
-		// Coupon Code Substep
-		add_action( 'fc_output_step_payment', array( $this, 'output_substep_coupon_codes' ), 10 );
-		add_filter( 'woocommerce_update_order_review_fragments', array( $this, 'add_coupon_codes_text_fragment' ), 10 );
+		// Checkout validation settings
+		add_filter( 'fc_js_settings', array( $this, 'add_js_settings' ), 10 );
 
 		// Prevent hiding coupon code field behind a link button, as it is implemented directly
 		add_filter( 'fc_hide_optional_fields_skip_list', array( $this, 'prevent_hide_optional_fields_coupon_code' ), 10 );
-	}
 
+		// Actions
+		add_action( 'wc_ajax_fc_add_coupon_code', array( $this, 'add_coupon_code' ), 10 );
+		add_action( 'wc_ajax_fc_remove_coupon_code', array( $this, 'remove_coupon_code' ), 10 );
 
+		// Integrated coupon code section at checkout
+		if ( 'yes' === get_option( 'fc_enable_checkout_coupon_codes', 'yes' ) ) {
+			// Checkout Coupon Notice
+			remove_action( 'woocommerce_before_checkout_form', 'woocommerce_checkout_coupon_form', 10 );
 
-	/**
-	 * Return Checkout Steps class instance.
-	 */
-	public function checkout_steps() {
-		return FluidCheckout_Steps::instance();
+			// Coupon Code Substep
+			add_action( 'fc_output_step_payment', array( $this, 'output_substep_coupon_codes' ), 10 );
+			add_filter( 'woocommerce_update_order_review_fragments', array( $this, 'add_coupon_codes_text_fragment' ), 10 );
+		}
 	}
 
 
@@ -62,6 +67,51 @@ class FluidCheckout_CouponCodes extends FluidCheckout {
 
 
 	/**
+	 * Register assets.
+	 */
+	public function register_assets() {
+		// Maybe load RTL file
+		$rtl_suffix = is_rtl() ? '-rtl' : '';
+
+		// Scripts
+		wp_register_script( 'fc-checkout-coupons', self::$directory_url . 'js/checkout-coupons'. self::$asset_version . '.js', array( 'jquery', 'require-bundle' ), NULL, true );
+		wp_add_inline_script( 'fc-checkout-coupons', 'window.addEventListener("load",function(){CheckoutCoupons.init(fcSettings.checkoutCoupons);})' );
+
+		// Styles
+		wp_register_style( 'fc-checkout-coupons', self::$directory_url . 'css/checkout-coupons'. $rtl_suffix . self::$asset_version . '.css', NULL, NULL );
+	}
+
+	/**
+	 * Enqueue assets.
+	 */
+	public function enqueue_assets() {
+		if ( is_checkout() && 'yes' === get_option( 'fc_enable_checkout_coupon_codes', 'yes' ) ) {
+			// Checkout steps scripts
+			wp_enqueue_script( 'fc-checkout-coupons' );
+			wp_enqueue_style( 'fc-checkout-coupons' );
+		}
+	}
+
+
+
+	/**
+	 * Add settings to the plugin settings JS object.
+	 *
+	 * @param   array  $settings  JS settings object of the plugin.
+	 */
+	public function add_js_settings( $settings ) {
+
+		$settings[ 'checkoutCoupons' ] = apply_filters( 'fc_checkout_coupons_script_settings', array(
+			'addCouponCodeNonce' => wp_create_nonce( 'fc-add-coupon-code' ),
+			'removeCouponCodeNonce' => wp_create_nonce( 'fc-remove-coupon-code' ),
+		) );
+
+		return $settings;
+	}
+
+
+
+	/**
 	 * Output coupon codes substep.
 	 *
 	 * @param   string  $step_id     Id of the step in which the substep will be rendered.
@@ -69,29 +119,33 @@ class FluidCheckout_CouponCodes extends FluidCheckout {
 	public function output_substep_coupon_codes( $step_id ) {
 		$substep_id = 'coupon_codes';
 		$substep_title = get_option( 'fc_display_coupon_code_section_title', 'no' ) === 'yes' ? apply_filters( 'fc_substep_coupon_codes_section_title', __( 'Coupon code', 'fluid-checkout' ) ) : null;
-		$this->checkout_steps()->output_substep_start_tag( $step_id, $substep_id, $substep_title );
+		FluidCheckout_Steps::instance()->output_substep_start_tag( $step_id, $substep_id, $substep_title );
 
+		$this->output_coupon_codes_messages_container();
 		$this->output_substep_text_coupon_codes();
 
-		$this->checkout_steps()->output_substep_fields_start_tag( $step_id, $substep_id, false );
+		FluidCheckout_Steps::instance()->output_substep_fields_start_tag( $step_id, $substep_id, false );
 		$this->output_substep_coupon_codes_fields();
-		$this->checkout_steps()->output_substep_fields_end_tag();
+		FluidCheckout_Steps::instance()->output_substep_fields_end_tag();
 
-		$this->checkout_steps()->output_substep_end_tag( $step_id, $substep_id, $substep_title, false );
+		FluidCheckout_Steps::instance()->output_substep_end_tag( $step_id, $substep_id, $substep_title, false );
 	}
 
 
 
 	/**
 	 * Output coupon codes fields.
+	 *
+	 * @param   array  $field_args               The coupon code field arguments. Will use default values if attributes are not passed in.
+	 * @param   array  $expansible_section_args  The attributes for the coupon code expansible section. Will use default values if attributes are not passed in.
 	 */
-	public function output_substep_coupon_codes_fields() {
+	public function output_substep_coupon_codes_fields( $field_args = array(), $expansible_section_args = array() ) {
 		$coupon_code_field_label = apply_filters( 'fc_coupon_code_field_label', __( 'Coupon code', 'fluid-checkout' ) );
 		$coupon_code_field_placeholder = apply_filters( 'fc_coupon_code_field_placeholder', __( 'Enter your code here', 'fluid-checkout' ) );
 		$coupon_code_button_label = apply_filters( 'fc_coupon_code_button_label', _x( 'Apply', 'Button label for applying coupon codes', 'fluid-checkout' ) );
 
 		$key = 'coupon_code';
-		$coupon_code_field_args = array(
+		$coupon_code_field_args = array_merge( array(
 			'required'                   => false,
 			'fc_skip_server_validation'  => true,
 			'class'                      => array( 'form-row-wide' ),
@@ -100,24 +154,23 @@ class FluidCheckout_CouponCodes extends FluidCheckout {
 				'aria-label'             => $coupon_code_field_label,
 				'data-autofocus'         => true,
 			),
-		);
+		), $field_args );
 
 		// Expansible section args
-		$coupon_code_expansible_args = array();
-		if ( apply_filters( 'fc_coupon_code_field_initially_expanded', false ) == true ) {
-			$coupon_code_expansible_args['initial_state'] = 'expanded';
-		}
+		$coupon_code_expansible_args = array_merge( array(
+			'initial_state' => true === apply_filters( 'fc_coupon_code_field_initially_expanded', false ) ? 'expanded' : 'collapsed',
+		), $expansible_section_args );
 
 		// Output coupon code field and button in an expansible form section
 		$coupon_code_toggle_label = get_option( 'fc_optional_fields_link_label_lowercase', 'yes' ) === 'yes' ? strtolower( $coupon_code_field_label ) : $coupon_code_field_label;
 		/* translators: %s: Form field label */
 		$coupon_code_toggle_label = apply_filters( 'fc_expansible_section_toggle_label_'.$key, sprintf( __( 'Add %s', 'fluid-checkout' ), $coupon_code_toggle_label ) );
-		$this->checkout_steps()->output_expansible_form_section_start_tag( $key, $coupon_code_toggle_label, $coupon_code_expansible_args );
+		FluidCheckout_Steps::instance()->output_expansible_form_section_start_tag( $key, $coupon_code_toggle_label, $coupon_code_expansible_args );
 		woocommerce_form_field( $key, $coupon_code_field_args );
 		?>
 		<button type="button" class="fc-coupon-code__apply <?php echo esc_attr( apply_filters( 'fc_coupon_code_apply_button_classes', 'button' ) ); ?>" data-apply-coupon-button><?php echo esc_html( $coupon_code_button_label ); ?></button>
 		<?php
-		$this->checkout_steps()->output_expansible_form_section_end_tag();
+		FluidCheckout_Steps::instance()->output_expansible_form_section_end_tag();
 	}
 
 
@@ -181,6 +234,97 @@ class FluidCheckout_CouponCodes extends FluidCheckout {
 	 */
 	public function output_substep_text_coupon_codes() {
 		echo $this->get_substep_text_coupon_codes();
+	}
+
+	/**
+	 * Output the coupon codes messages container element.
+	 */
+	public function output_coupon_codes_messages_container() {
+		echo '<div class="fc-coupon-code-messages"></div>';
+	}
+
+
+
+	/**
+	 * AJAX Add a coupon code to the cart.
+	 * @see WC_AJAX::apply_coupon
+	 */
+	public function add_coupon_code() {
+		check_ajax_referer( 'fc-add-coupon-code', 'security' );
+
+		$coupon_code = sanitize_text_field( wp_unslash( $_REQUEST['coupon_code'] ) );
+
+		if ( ! empty( $coupon_code ) ) {
+			WC()->cart->add_discount( wc_format_coupon_code( $coupon_code ) );
+
+			// Intercept notices to avoid them being displayed on other pages
+			ob_start();
+			wc_print_notices();
+			$message = ob_get_clean();
+			
+			wp_send_json(
+				array(
+					'result'           => false === strpos( $message, 'woocommerce-error' ) ? 'success' : 'error',
+					'coupon_code'      => $coupon_code,
+					'message'          => $message,
+				)
+			);
+		}
+		else {
+			wp_send_json(
+				array(
+					'result'           => 'error',
+					'coupon_code'      => $coupon_code,
+					'error_slug'       => 'coupon_code_does_not_exist',
+					'message'          => WC_Coupon::get_generic_coupon_error( WC_Coupon::E_WC_COUPON_PLEASE_ENTER ),
+				)
+			);
+		}
+	}
+
+	/**
+	 * AJAX Remove a coupon code from the cart.
+	 * @see WC_AJAX::remove_coupon
+	 */
+	public function remove_coupon_code() {
+		check_ajax_referer( 'fc-remove-coupon-code', 'security' );
+
+		$coupon_code = sanitize_text_field( wp_unslash( $_REQUEST['coupon_code'] ) );
+
+		if ( ! empty( $coupon_code ) ) {
+			WC()->cart->remove_coupon( wc_format_coupon_code( $coupon_code ) );
+			wc_add_notice( __( 'Coupon has been removed.', 'woocommerce' ) );
+
+			// Intercept notices to avoid them being displayed on other pages
+			ob_start();
+			wc_print_notices();
+			$message = ob_get_clean();
+			
+			wp_send_json(
+				array(
+					'result'           => false === strpos( $message, 'woocommerce-error' ) ? 'success' : 'error',
+					'coupon_code'      => $coupon_code,
+					'message'          => $message,
+				)
+			);
+		}
+		else {
+			wc_add_notice( __( 'Sorry there was a problem removing this coupon.', 'woocommerce' ), 'error' );
+
+			// Intercept notices to avoid them being displayed on other pages
+			ob_start();
+			wc_print_notices();
+			$message = ob_get_clean();
+
+			wp_send_json(
+				array(
+					'result'           => 'error',
+					'coupon_code'      => $coupon_code,
+					'error_slug'       => 'coupon_code_not_provided',
+					'message'          => $message,
+				)
+			);
+		}
 	}
 
 }