+++ inc/checkout-validation.php
@@ -26,10 +26,11 @@ class FluidCheckout_Validation extends FluidCheckout {
 		add_filter( 'body_class', array( $this, 'add_body_class' ) );
 
 		// Enqueue assets
+		add_action( 'wp_enqueue_scripts', array( $this, 'register_assets' ), 5 );
 		add_action( 'wp_enqueue_scripts', array( $this, 'enqueue_assets' ), 10 );
 
 		// Checkout validation settings
-		add_filter( 'fc_js_settings', array( $this, 'add_checkout_validation_js_settings' ), 10 );
+		add_filter( 'fc_js_settings', array( $this, 'add_js_settings' ), 10 );
 
 		// Mailcheck validation
 		add_filter( 'fc_checkout_field_args' , array( $this, 'change_checkout_email_field_args' ), 10 );
@@ -59,6 +60,21 @@ class FluidCheckout_Validation extends FluidCheckout {
 
 
 	/**
+	 * Register assets.
+	 */
+	public function register_assets() {
+		// Maybe load RTL file
+		$rtl_suffix = is_rtl() ? '-rtl' : '';
+		
+		// Styles
+		wp_register_style( 'fc-checkout-validation', self::$directory_url . 'css/checkout-validation'. $rtl_suffix . self::$asset_version . '.css', NULL, NULL );
+
+		// Checkout steps scripts
+		wp_register_script( 'fc-checkout-validation', self::$directory_url . 'js/checkout-validation'. self::$asset_version . '.js', array( 'jquery', 'wc-checkout' ), NULL, true );
+		wp_add_inline_script( 'fc-checkout-validation', 'window.addEventListener("load",function(){CheckoutValidation.init(fcSettings.checkoutValidation);})' );
+	}
+
+	/**
 	 * Enqueue scripts.
 	 */
 	public function enqueue_assets() {
@@ -66,27 +82,21 @@ class FluidCheckout_Validation extends FluidCheckout {
 		if( ! function_exists( 'is_checkout' ) || ! is_checkout() ){ return; }
 
 		// Styles
-		if ( is_rtl() ) {
-			wp_enqueue_style( 'fc-checkout-validation', self::$directory_url . 'css/checkout-validation-rtl'. self::$asset_version . '.css', NULL, NULL );
-		}
-		else {
-			wp_enqueue_style( 'fc-checkout-validation', self::$directory_url . 'css/checkout-validation'. self::$asset_version . '.css', NULL, NULL );
-		}
+		wp_enqueue_style( 'fc-checkout-validation' );
 
 		// Checkout steps scripts
-		wp_enqueue_script( 'fc-checkout-validation', self::$directory_url . 'js/checkout-validation'. self::$asset_version . '.js', array( 'jquery', 'wc-checkout' ), NULL, true );
-		wp_add_inline_script( 'fc-checkout-validation', 'window.addEventListener("load",function(){CheckoutValidation.init(fcSettings.checkoutValidation);})' );
+		wp_enqueue_script( 'fc-checkout-validation' );
 	}
 
 
 
 
 	/**
-	 * Add Checkout Validation settings to the plugin settings JS object.
+	 * Add settings to the plugin settings JS object.
 	 *
 	 * @param   array  $settings  JS settings object of the plugin.
 	 */
-	public function add_checkout_validation_js_settings( $settings ) {
+	public function add_js_settings( $settings ) {
 
 		$settings[ 'checkoutValidation' ] = apply_filters( 'fc_checkout_validation_script_settings', array(
 			'formRowSelector'                    => '.form-row, .shipping-method__package',